modules/BMP280Module/requirements.txt
azure-iot-device~=2.0.0
cffi
smbus-cffi==0.5.1